Fix integrating of alt and title attributes
The alt attribute was only integrated in the img tag if it was empty. This
is now fixed. Also if the alt attribute is given directly in the view helper
is considered now.

Additionally the integration of the title attribute is now also fixed. If
the title attribute is given directly in the view helper is considered
now.