[TASK] Apply maxWidth for srcset's w/o explicit ratio requirements
Use maxWidth instead of width as ProcessingInstructions to ensure
small images won't be upscaled to larger srcset demands,
when maxWidth is used instead of width on the image tag.

A 400px source image should not be scaled to 600px, even
if "600" is defined in `variants` when maxWidth is used.
Instead of upscaling, an image, no larger than the original one
(and if there is no cropping defined,
that would be the original file), is now be returned.

Note that this change causes the hash of ProcessingInstructions
to change for cases where variants="…" is used when maxWidth
is used without width/height, thus processed images will need to
be regenerated – but that's unlikely to be the case, due to the
previous behavior, that ignored maxWidth.